Separated `api-` handler from another handlers
