+++ server/handlers/api-browsers.js
@@ -6,7 +6,7 @@ import getBrowsers, {
 } from '../lib/get-browsers.js'
 import { sendResponseAPI } from '../lib/send-response.js'
 
-export default async function handleBrowsers(req, res) {
+export default async function handleAPIBrowsers(req, res) {
   let { searchParams: params } = new URL(req.url, `http://${req.headers.host}/`)
 
   let query = params.get('q') || QUERY_DEFAULTS
