+++ server/handlers/static.js
